refactor(evaluator): use static analysis

+++ crates/jrsonnet-evaluator/src/typed/conversions.rs
@@ -637,7 +637,7 @@ impl FromUntyped for IBytes {
 		}
 		<Self as Typed>::TYPE.check(&value)?;
 		// Any::downcast_ref::<ByteArray>(&a);
-		let mut out = Vec::with_capacity(a.len());
+		let mut out = Vec::with_capacity(a.len() as usize);
 		for e in a.iter() {
 			let r = e?;
 			out.push(u8::from_untyped(r)?);
